Pop Scope Changed in chat_screen.deart
Hi Brother,

I watched your video tutorial on the Chat App and used your chat UI. However, I encountered an issue with the pop scope. In release mode, trying to go back from the chat screen to the home screen using the back button didn't work, and the back button in the app bar was also unresponsive. I managed to fix this problem in my app.